Refactor: Introduce service architecture and server mode
This commit refactors Clovie to use a service-oriented architecture and introduces a new server mode for building full Express applications. The documentation and tests have been updated to reflect these changes.
