+++ lib/Cache.js
@@ -37,7 +37,7 @@ export class Cache extends ServiceProvider {
   actions(useContext) {
     const { stable, config } = useContext();
     
-    return {
+    const actions = {
       getFileHash: (filePath) => {
         try {
           if (!fs.existsSync(filePath)) return null;
@@ -49,9 +49,15 @@ export class Cache extends ServiceProvider {
       },
 
       hasChanged: (filePath) => {
-        const currentHash = this.getFileHash(filePath);
-        const cache = stable.get(['cache']);
-        if (!cache) {
+        const currentHash = actions.getFileHash(filePath);
+        let cache = stable.get(['cache']);
+        if (!cache || !cache.files) {
+          cache = {
+            files: {},
+            lastBuild: null,
+            stats: { totalBuilds: 0, averageBuildTime: 0 }
+          };
+          stable.set(['cache'], cache);
           return true; // If no cache exists, consider it changed
         }
         const cachedHash = cache.files[filePath];
@@ -65,7 +71,7 @@ export class Cache extends ServiceProvider {
       },
 
       getChangedFiles: (files) => {
-        return files.filter(file => this.hasChanged(file));
+        return files.filter(file => actions.hasChanged(file));
       },
 
       markBuilt: () => {
@@ -153,6 +159,8 @@ export class Cache extends ServiceProvider {
         }
       }
     };
+    
+    return actions;
   }
 
   #migrateFromFile(stable) {
