+++ lib/File.js
@@ -50,14 +50,15 @@ export class File extends ServiceProvider {
               fs.mkdirSync(dir, { recursive: true });
             }
           
-            // Check if the value is a Buffer (binary data) or string
-            const options = Buffer.isBuffer(value) ? {} : { encoding: 'utf8' };
-            
-            fs.writeFileSync(dest, value, options, err => {
-              if (err) {
-                console.log(err);
-              }
-            });
+            try {
+              // Check if the value is a Buffer (binary data) or string
+              const options = Buffer.isBuffer(value) ? {} : { encoding: 'utf8' };
+              
+              fs.writeFileSync(dest, value, options);
+            } catch (err) {
+              console.error(`Error writing file ${dest}:`, err);
+              throw err;
+            }
           }
         }
