+++ lib/Server.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import path from 'path';
+import fs from 'fs';
 import { createServer } from 'http';
 import { Server as SocketIOServer } from 'socket.io';
 import { ServiceProvider } from '@brickworks/engine';
@@ -55,23 +56,32 @@ export class Server extends ServiceProvider {
       await routes.setupServerRoutes(this.#app);
     } else {
       console.log('📁 Initializing static file server type...');
-      // In static mode, serve static files from output directory
-      this.#app.use(express.static(config.outputDir));
-      
-      // Explicitly handle root path to serve index.html
-      this.#app.get('/', (req, res) => {
-        res.sendFile(path.join(config.outputDir, 'index.html'));
-      });
-      
-      // Handle paths without trailing slash by redirecting
-      this.#app.get('*', (req, res, next) => {
-        if (!req.path.endsWith('/') && !req.path.includes('.')) {
-          // Redirect paths without trailing slash to include it
-          res.redirect(req.path + '/');
-          return;
-        }
-        next();
-      });
+      // In static mode, serve static files from output directory (if provided)
+      if (config.outputDir && fs.existsSync(config.outputDir)) {
+        this.#app.use(express.static(config.outputDir));
+        
+        // Explicitly handle root path to serve index.html
+        this.#app.get('/', (req, res) => {
+          const indexPath = path.join(config.outputDir, 'index.html');
+          if (fs.existsSync(indexPath)) {
+            res.sendFile(indexPath);
+          } else {
+            res.status(404).send('index.html not found');
+          }
+        });
+        
+        // Handle paths without trailing slash by redirecting
+        this.#app.get('*', (req, res, next) => {
+          if (!req.path.endsWith('/') && !req.path.includes('.')) {
+            // Redirect paths without trailing slash to include it
+            res.redirect(req.path + '/');
+            return;
+          }
+          next();
+        });
+      } else {
+        console.log('⚠️  No output directory configured or it does not exist, serving basic routes only');
+      }
     }
     
     // Setup development features
